[Change] Move driveFiles, storageQuota, aerialFolder to layout server load function
